+++ plistsync/core/crdt/__init__.py
@@ -1,6 +1,6 @@
 """Minimal Conflict Free Replicatable Data Types implementations."""
 
-from plistsync.crdt.fugue import DeleteOp, Fugue, InsertOp, InsertPos
+from .fugue import DeleteOp, Fugue, InsertOp, InsertPos
 
 __all__ = [
     "DeleteOp",
